Change the arrayUtils.ts in members to common modules.
